styles : icon fixes
