+++ apps/website/components/molecules/mint-card/assets/token-filled.tsx
