+++ libs/assets/src/index.ts
@@ -2,4 +2,4 @@ export * from './lib/coinbase-wallet';
 export * from './lib/gateway';
 export * from './lib/metamask';
 export * from './lib/walletconnect';
-export * from './lib/token-filled'
+
