+++ libs/assets/src/lib/gateway.tsx
@@ -17,24 +17,4 @@ export function GatewayIcon(props: SvgIconProps) {
   );
 }
 
-export function GatewayGrayIcon(props: SvgIconProps) {
-  return (
-    <SvgIcon viewBox="0 0 28 28" {...props}>
-      <path
-        d="M13.8679 0C12.0468 -2.19239e-08 10.2434 0.362121 8.56091 1.06569C6.87837 1.76925 5.34958 2.80048 4.06183 4.1005C2.77407 5.40052 1.75256 6.94387 1.05563 8.64243C0.358705 10.341 -5.78884e-08 12.1615 0 14C5.78884e-08 15.8385 0.358705 17.659 1.05563 19.3576C1.75256 21.0561 2.77407 22.5995 4.06183 23.8995C5.34958 25.1995 6.87837 26.2307 8.56091 26.9343C10.2434 27.6379 12.0468 28 13.8679 28L13.8679 0Z"
-        fill="white"
-        fillOpacity="0.3"
-      />
-      <path
-        d="M15.9812 0L16.9058 1.20248e-06C23.033 9.17164e-06 28.0001 5.01442 28.0001 11.2L15.9812 0Z"
-        fill="white"
-        fillOpacity="0.3"
-      />
-      <path
-        d="M28.0001 28V18.0098C28.0001 14.7643 25.3096 12.1333 21.9906 12.1333C18.6717 12.1333 15.9812 14.7643 15.9812 18.0098V28H28.0001Z"
-        fill="white"
-        fillOpacity="0.3"
-      />
-    </SvgIcon>
-  );
-}
+
